+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.4.1
+- Reformatted code to 80 chars wide
+
+## 1.4.0
+- Upgraded bilocator to 1.6.0
+- Cleanup of renaming of View to ViewModel and StatelessView to StatelessViewWidget
+
 ## 1.3.0
 - Renamed View to ViewWidget because of a name collision on the Flutter beta channel.
 - Renamed StatelessView to StatelessViewWidget to be consistent with ViewWidget renaming.